chat_app_backend/requirements.txt
Django
djangorestframework
channels
channels-redis